Change Embedder import to on-demand
Currently the file imports Embedder even if it's not in use. This is an issue if we are using an external embedder and do not have PyTorch/Tensorflow installed. It also bloats up the process memory unnecessarily.
This commit changes the import to happen only if the Embedder is required.